backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
sqlparse==0.5.3
django-cors-headers==4.3.1
Pillow==10.0.1
